rename test to be more specific
