+++ corehq/apps/userreports/tests/__init__.py
@@ -1,4 +1,4 @@
-from .test_config import *
+from .test_indicator_config import *
 from .test_filters import *
 from .test_indicators import *
 from .test_pillow import *
